# IGNITE-22501 Sql. Batching DDL statement for scripts