--- a/modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/DdlBatchingBenchmark.java
+++ b/modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/DdlBatchingBenchmark.java
@@ -114,7 +114,7 @@ public class DdlBatchingBenchmark extends AbstractMultiNodeBenchmark {
     }
 
     @Override
-    protected void createTableOnStartup() {
+    protected void createTablesOnStartup() {
         // no-op
     }