--- a/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItIndexAvailabilityTest.java
+++ b/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItIndexAvailabilityTest.java
@@ -43,7 +43,7 @@ class ItIndexAvailabilityTest extends BaseSqlIntegrationTest {
 
     @BeforeEach
     void setUp() {
-        sql(String.format("CREATE TABLE IF NOT EXISTS %s (key BIGINT PRIMARY KEY, val VARCHAR)", TABLE_NAME));
+        sql(createTableStatement(TABLE_NAME));
     }
 
     @AfterEach
@@ -58,5 +58,31 @@ class ItIndexAvailabilityTest extends BaseSqlIntegrationTest {
         assertQuery(format("SELECT * FROM {} WHERE val = 'test'", TABLE_NAME))
                 .matches(containsIndexScan(SqlCommon.DEFAULT_SCHEMA_NAME, TABLE_NAME, INDEX_NAME))
                 .check();
+    }
+
+    /**
+     * Similar to {@link #indexIsQueriableRightAfterCreationFutureCompletes()}, but index is created as second statement
+     * of script, and first item is conditional statement that won't be applied.
+     */
+    @Test
+    void indexIsQueriableRightAfterCreationFutureCompletes2() {
+        // Make sure table with required name already exists.
+        assertQuery("SELECT COUNT(*) FROM system.tables WHERE name = ?")
+                .withParam(TABLE_NAME)
+                .returns(1L)
+                .check();
+
+        CLUSTER.aliveNode().sql().executeScript(
+                createTableStatement(TABLE_NAME) + ";"
+                + String.format("CREATE INDEX %s ON %s (val)", INDEX_NAME, TABLE_NAME)
+        );
+
+        assertQuery(format("SELECT * FROM {} WHERE val = 'test'", TABLE_NAME))
+                .matches(containsIndexScan(SqlCommon.DEFAULT_SCHEMA_NAME, TABLE_NAME, INDEX_NAME))
+                .check();
+    }
+
+    private static String createTableStatement(String tableName) {
+        return String.format("CREATE TABLE IF NOT EXISTS %s (key BIGINT PRIMARY KEY, val VARCHAR)", tableName);
     }
 }