--- a/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItSqlMultiStatementTest.java
+++ b/modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItSqlMultiStatementTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.ignite.internal.sql.engine;
 
+import static org.apache.ignite.internal.sql.engine.util.QueryChecker.containsIndexScan;
 import static org.apache.ignite.internal.sql.engine.util.SqlTestUtils.assertThrowsSqlException;
 import static org.apache.ignite.internal.testframework.IgniteTestUtils.assertThrows;
 import static org.apache.ignite.internal.testframework.IgniteTestUtils.await;
@@ -235,5 +236,32 @@ public class ItSqlMultiStatementTest extends BaseSqlMultiStatementTest {
         assertThat(batch.items().get(0).get(0), is(tableSize));
 
         iterateThroughResultsAndCloseThem(cursor);
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    void indexesAvailableAfterScriptExecutionAndBuiltProperly() {
+        long tableSize = 1_000;
+
+        @SuppressWarnings("ConcatenationWithEmptyString")
+        String script = ""
+                + "CREATE TABLE integers (id INT PRIMARY KEY, val_1 INT, val_2 INT);"
+                + "CREATE INDEX integers_val_1_ind ON integers(val_1);"
+                + "INSERT INTO integers SELECT x, x, x FROM system_range(1, " + tableSize + ");"
+                + "CREATE INDEX integers_val_2_ind ON integers(val_2);";
+
+        AsyncSqlCursor<InternalSqlRow> cursor = runScript(script);
+
+        iterateThroughResultsAndCloseThem(cursor);
+
+        assertQuery("SELECT /*+ FORCE_INDEX(INTEGERS_VAL_1_IND) */ COUNT(*) FROM integers WHERE val_1 > 0")
+                .matches(containsIndexScan("PUBLIC", "INTEGERS", "INTEGERS_VAL_1_IND"))
+                .returns(tableSize)
+                .check();
+
+        assertQuery("SELECT /*+ FORCE_INDEX(INTEGERS_VAL_2_IND) */ COUNT(*) FROM integers WHERE val_2 > 0")
+                .matches(containsIndexScan("PUBLIC", "INTEGERS", "INTEGERS_VAL_2_IND"))
+                .returns(tableSize)
+                .check();
     }
 }