… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/CurrentTimeProvider.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/fsm/QueryExecutor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/fsm/ScriptItemPrepareProgram.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/NodeLeaveHandler.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/DdlBatchingTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/ExecutionServiceImplTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/ddl/DdlCommandHandlerExceptionHandlingTest.java
… modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/framework/TestNode.java