https://issues.apache.org/jira/browse/IGNITE-22501

Result of benchmark introduced in this patch on my laptop as follow:
```
Before
Benchmark                  (fsync)  (numberOfSecondaryIndexesPerTable)  (numberOfTables)  Mode  Cnt   Score   Error  Units
DdlBatchingBenchmark.test    false                                   0                 1  avgt    7   2.016 ± 0.005   s/op
DdlBatchingBenchmark.test    false                                   0                 2  avgt    7   4.026 ± 0.007   s/op
DdlBatchingBenchmark.test    false                                   0                 3  avgt    7   6.040 ± 0.007   s/op
DdlBatchingBenchmark.test    false                                   2                 1  avgt    7   7.999 ± 0.023   s/op
DdlBatchingBenchmark.test    false                                   2                 2  avgt    7  15.998 ± 0.026   s/op
DdlBatchingBenchmark.test    false                                   2                 3  avgt    7  23.999 ± 0.017   s/op
DdlBatchingBenchmark.test    false                                   4                 1  avgt    7  13.998 ± 0.016   s/op
DdlBatchingBenchmark.test    false                                   4                 2  avgt    7  27.999 ± 0.010   s/op
DdlBatchingBenchmark.test    false                                   4                 3  avgt    7  41.999 ± 0.035   s/op

After
Benchmark                  (fsync)  (numberOfSecondaryIndexesPerTable)  (numberOfTables)  Mode  Cnt  Score   Error  Units
DdlBatchingBenchmark.test    false                                   0                 1  avgt    7  1.008 ± 0.004   s/op
DdlBatchingBenchmark.test    false                                   0                 2  avgt    7  1.008 ± 0.003   s/op
DdlBatchingBenchmark.test    false                                   0                 3  avgt    7  1.009 ± 0.006   s/op
DdlBatchingBenchmark.test    false                                   2                 1  avgt    7  1.010 ± 0.006   s/op
DdlBatchingBenchmark.test    false                                   2                 2  avgt    7  1.009 ± 0.006   s/op
DdlBatchingBenchmark.test    false                                   2                 3  avgt    7  1.010 ± 0.004   s/op
DdlBatchingBenchmark.test    false                                   4                 1  avgt    7  1.012 ± 0.007   s/op
DdlBatchingBenchmark.test    false                                   4                 2  avgt    7  1.011 ± 0.007   s/op
DdlBatchingBenchmark.test    false                                   4                 3  avgt    7  1.010 ± 0.003   s/op
```
-------------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)